refactor(render): move rendering to separate module
- move renderer to new `NLockRenderer` struct
- move rendering enums into `util` module
- clean up `surface` module
